refactor(sandbox): extract sandbox package
