+++ packages/server/package.json
@@ -34,6 +34,7 @@
         "@nangohq/nango-yaml": "file:../nango-yaml",
         "@nangohq/providers": "file:.../providers",
         "@nangohq/records": "file:../records",
+        "@nangohq/sandbox": "file:../sandbox",
         "@nangohq/shared": "file:../shared",
         "@nangohq/utils": "file:../utils",
         "@nangohq/webhooks": "file:../webhooks",
@@ -45,7 +46,6 @@
         "cors": "2.8.5",
         "dd-trace": "5.52.0",
         "dotenv": "16.5.0",
-        "e2b": "2.18.0",
         "exponential-backoff": "3.1.1",
         "express": "5.1.0",
         "express-session": "1.18.2",
