+++ packages/shared/lib/index.ts
@@ -13,7 +13,6 @@ import flowService from './services/flow.service.js';
 import hmacService from './services/hmac.service.js';
 import { errorNotificationService } from './services/notification/error.service.js';
 import { SlackService, generateSlackConnectionId } from './services/notification/slack.service.js';
-import sandboxApiKeyService from './services/sandbox-api-key.service.js';
 import secretService from './services/secret.service.js';
 import sharedCredentialsService from './services/shared-credentials.service.js';
 import syncManager, { syncCommandToOperation } from './services/sync/manager.service.js';
@@ -29,7 +28,6 @@ export * as jwtClient from './auth/jwt.js';
 export * as signatureClient from './auth/signature.js';
 export * from './services/connections/credentials/refresh.js';
 export * from './services/function.service.js';
-export * from './services/functions/dryruns.service.js';
 export * from './services/on-event-scripts.service.js';
 export * from './services/sync/sync.service.js';
 export * from './services/sync/job.service.js';
@@ -50,20 +48,6 @@ export * from './services/checkpoints/checkpoints.js';
 export * from './services/shared-credentials.service.js';
 export * as connectUISettingsService from './services/connect-ui-settings.service.js';
 export { deployTemplate, upgradeTemplate } from './services/deploy/template.js';
-export {
-    buildSandboxApiKeyScopes,
-    createSandboxApiKeyToken,
-    createSandboxSigningSecret,
-    decryptSandboxSigningSecret,
-    encryptSandboxSigningSecret,
-    isSandboxApiKey,
-    parseSandboxApiKeyToken,
-    sandboxApiKeyAudience,
-    sandboxApiKeyBaseScopes,
-    sandboxApiKeyPrefix,
-    sandboxApiKeyPurposes,
-    verifySandboxApiKeyToken
-} from './services/sandbox-api-key.service.js';
 
 export * as oauth2Client from './clients/oauth2.client.js';
 export * as mcpClient from './clients/mcp.client.js';
@@ -100,7 +84,6 @@ export {
     pbkdf2,
     providerClientManager,
     remoteFileService,
-    sandboxApiKeyService,
     secretService,
     seeders,
     sharedCredentialsService,
